requirements.txt
# Pixel2Mesh requirements
# Target: Python 3.12 | PyTorch 2.5.1 | CUDA 12.6 (cu124 wheels)
#
# Install PyTorch first (not listed here — handled in Dockerfile):
#   pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1 \
#       --index-url https://download.pytorch.org/whl/cu124

# Core
numpy
scipy>=1.3
scikit-image>=0.15
opencv-python>=4.1
Pillow

# Config & logging
easydict
pyyaml
tensorboardX

# Mesh processing
trimesh
shapely

# Optional: Design C (FaceScape)
# nvidia-dali-cuda120   # uncomment for DALI GPU data loading
